STREAMLIT/requirements.txt
streamlit
requests
python-dotenv
jsons
cryptography
pandas
# opencv-python
scikit-learn
face_recognition
 
